src/requirements.txt
pyrogram
pytgcrypto
tortoise-orm
aiohttp